+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.0.3 / Unreleased
 
 - #246 Add dir_browser/htdocs folder setup.cfg (for install with `pip install  .`)
+- #268 Use relative paths to support reverse proxies
 
 ## 4.0.2 / 2022-08-01